feat: improve MCP attribution and FTS token-aware search
- Propagate upstream agent/session across bridge via _so_agent/_so_session_id
- Use forwarded identity in so_read/so_search/so_status event attribution
- Add centralized token counting with tiktoken-rs (cl100k_base)
- Store per-file token_count in graph schema and index/sync paths
- Update so_search stats to use matched file token sums
- Add FTS query normalization (strip specials, boolean guard, prefix terms)
- Rank search results with bm25 and include score in output
- Move watch/unwatch control to dedicated ctrl socket path handling
- Improve daemon watch logs with explicit initial index start line
- Fix OpenCode plugin template command variables (cmd/dir)
- Keep FTS table data intact across DB opens (no drop-on-open)

+++ src/core/graph/mod.rs
@@ -53,8 +53,8 @@ pub fn sync_project(project_path: &str) -> Result<String, String> {
 }
 
 /// Searches indexed graph content for a query.
-/// Returns formatted results and total on-disk char count of matched files.
-pub fn search_project_with_stats(project_path: &str, query: &str, limit: usize) -> Result<(String, usize), String> {
+/// Returns formatted results and total token count of matched files.
+pub fn search_project_with_stats(project_path: &str, query: &str, limit: usize) -> Result<(String, i64), String> {
     let project_root = validate_project_root(project_path)?;
     if !GraphDb::exists(&project_root) {
         let db_path = db::graph_db_path(&project_root);
